+++ libs/common/src/lib/ai-models/ai-models.component.ts
@@ -21,6 +21,7 @@ import { LearningConfigurations, WritableKnowledgeBox } from '@nuclia/core';
 import { InfoCardComponent } from '@nuclia/sistema';
 import { removeDeprecatedModels } from './ai-models.utils';
 import { StandaloneService } from '../services';
+import { ExtractionComponent } from './extraction/extraction.component';
 
 @Component({
   selector: 'stf-ai-models',
@@ -38,6 +39,7 @@ import { StandaloneService } from '../services';
     SummarizationComponent,
     SemanticModelComponent,
     AnonymizationComponent,
+    ExtractionComponent,
     InfoCardComponent,
     UnauthorizedFeatureDirective,
   ],
@@ -48,7 +50,8 @@ import { StandaloneService } from '../services';
 export class AiModelsComponent implements OnInit {
   private unsubscribeAll = new Subject<void>();
 
-  selectedTab: 'anonymization' | 'answer-generation' | 'semantic-model' | 'summarization' = 'answer-generation';
+  selectedTab: 'anonymization' | 'answer-generation' | 'semantic-model' | 'summarization' | 'extraction' =
+    'answer-generation';
 
   kb?: WritableKnowledgeBox;
   learningConfigurations?: LearningConfigurations;
@@ -94,7 +97,7 @@ export class AiModelsComponent implements OnInit {
       });
   }
 
-  selectTab(tab: 'anonymization' | 'answer-generation' | 'semantic-model' | 'summarization'): void {
+  selectTab(tab: 'anonymization' | 'answer-generation' | 'semantic-model' | 'summarization' | 'extraction'): void {
     if (tab === 'summarization') {
       this.isSummarizationAuthorized
         .pipe(