+++ libs/sdk-core/src/lib/db/kb/kb.ts
@@ -17,6 +17,8 @@ import {
   Counters,
   Entities,
   EntitiesGroup,
+  ExtractConfig,
+  ExtractStrategies,
   FullKbUser,
   IKnowledgeBox,
   IKnowledgeBoxCreation,
@@ -1149,5 +1151,36 @@ export class WritableKnowledgeBox extends KnowledgeBox implements IWritableKnowl
    */
   removeVectorset(model: string): Observable<void> {
     return this.nuclia.rest.delete(`/kb/${this.id}/vectorsets/${model}`);
+  }
+
+  getExtractStrategies(): Observable<ExtractStrategies> {
+    return of({
+      test1: {
+        vllm_config: {
+          llm: { generative_model: 'gemini-1-5-flash-vision', generative_provider: 'google' },
+          rules: ['ignore titles'],
+        },
+        ai_tables: { llm: undefined },
+        split: { max_paragraph: 200 },
+      },
+      test2: {
+        ai_tables: {
+          llm: { generative_model: 'gemini-1-5-flash-vision', generative_provider: 'google' },
+        },
+        split: { max_paragraph: 220 },
+      },
+      test3: {},
+    } as any);
+    //return this.nuclia.rest.get<ExtractStrategies>('/extract_strategies');
+  }
+
+  createExtractStrategy(id: string, config: ExtractConfig): Observable<void> {
+    return of(undefined);
+    //return this.nuclia.rest.post<void>('/extract_strategies', config);
+  }
+
+  deleteExtractStrategy(id: string): Observable<void> {
+    return of(undefined);
+    //return this.nuclia.rest.delete(`/extract_strategies/${id}`);
   }
 }